Start merging PdoAdapter into AbstractAdapter
Because all adapters extend PdoAdapter, the intermediary class just adds
complexity and indirection. Merging the two classes together shows the
accumulation of a deity that could benefit from some decomposition.
